Updating shard functionality

+++ solr/scripts/remove_shard.sh
@@ -10,35 +10,40 @@ shard_name=$1
 shard_port=$2
 shard="$1-$2"
 
-# delete index directory /esg/solr-index/<host>-<port>/
-if [ -d "/esg/solr-index/${shard}" ]; then
-  echo "Deleting shard index directory: /esg/solr-index/${shard}"
-  rm -rf /esg/solr-index/${shard}
-fi
+# do not ever remove master or slave shards
+if ! [[ $shard_name == 'master' || $shard_name == 'slave' ]]; then
 
-# remove solr-home directory /usr/local/solr-home/<host>-<port>/
-if [ -d "$SOLR_HOME/${shard}" ]; then
-  echo "Deleting shard home directory $SOLR_HOME/${shard}"
-  rm -rf $SOLR_HOME/${shard}
-fi
-
-# remove supervisor configuration 
-supervisord_config="/etc/supervisor/conf.d/supervisord.solr_${shard_name}_${shard_port}.conf"
-if [ -f "${supervisord_config}" ]; then
-  echo "Deleting supervisord configuration: $supervisord_config"
-  rm $supervisord_config
-
-  # remove this program to the solr group
-  if grep -q ${shard_name} /etc/supervisor/conf.d/supervisord.solr.conf; then
-     echo "Removing $shard from supervisor group 'solr'"
-     sed -i '/^programs=/ s/,'${shard_name}'_'${shard_port}'//' /etc/supervisor/conf.d/supervisord.solr.conf
+  # delete index directory /esg/solr-index/<host>-<port>/
+  if [ -d "/esg/solr-index/${shard}" ]; then
+    echo "Deleting shard index directory: /esg/solr-index/${shard}"
+    rm -rf /esg/solr-index/${shard}
   fi
 
-fi
+  # remove solr-home directory /usr/local/solr-home/<host>-<port>/
+  if [ -d "$SOLR_HOME/${shard}" ]; then
+    echo "Deleting shard home directory $SOLR_HOME/${shard}"
+    rm -rf $SOLR_HOME/${shard}
+  fi
 
-# remove shard from list queried by ESGF search application
-shards_file="/esg/config/esgf_shards_static.xml"
-if grep -q ${shard_port} ${shards_file} ; then
-   echo "Removing shard from ${shards_file}"
-   sed -i '/localhost:'${shard_port}'/d' ${shards_file}
+  # remove supervisor configuration 
+  supervisord_config="/etc/supervisor/conf.d/supervisord.solr_${shard_name}_${shard_port}.conf"
+  if [ -f "${supervisord_config}" ]; then
+    echo "Deleting supervisord configuration: $supervisord_config"
+    rm $supervisord_config
+
+    # remove this program to the solr group
+    if grep -q ${shard_name} /etc/supervisor/conf.d/supervisord.solr.conf; then
+      echo "Removing $shard from supervisor group 'solr'"
+      sed -i '/^programs=/ s/,'${shard_name}'_'${shard_port}'//' /etc/supervisor/conf.d/supervisord.solr.conf
+    fi
+
+  fi
+
+  # remove shard from list queried by ESGF search application
+  shards_file="/esg/config/esgf_shards_static.xml"
+  if grep -q ${shard_port} ${shards_file} ; then
+    echo "Removing shard from ${shards_file}"
+    sed -i '/localhost:'${shard_port}'/d' ${shards_file}
+  fi
+
 fi
